refactor(@formatjs/editor): use native Playwright visual tests
