+++ packages/editor/vrt/app.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import App from '../index'
+
+ReactDOM.render(<App />, document.getElementById('editor-root'))
