+++ packages/editor/vrt/index.html
@@ -1,0 +1,11 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Visual test fixture</title>
+  </head>
+  <body>
+    <div id="editor-root"></div>
+    <script type="module" src="/app.tsx"></script>
+  </body>
+</html>
