+++ packages/editor/vrt/package.json
@@ -7,19 +7,18 @@
     "@material-ui/core": "4.12.4",
     "@material-ui/icons": "4.11.3",
     "@material-ui/lab": "4.0.0-alpha.61",
+    "@playwright/test": "1.62.0",
     "@types/node": "24.10.1",
     "@types/react": "17.0.90",
     "@types/react-dom": "17.0.26",
     "@vitejs/plugin-react": "6.0.1",
-    "@vitest/browser-playwright": "4.1.6",
     "playwright": "1.62.0",
     "playwright-core": "1.62.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-intl": "6.8.9",
     "typescript": "5.9.3",
-    "vite": "8.1.0",
-    "vitest": "4.1.6"
+    "vite": "8.1.0"
   },
   "packageManager": "pnpm@11.17.0"
 }
