+++ packages/editor/vrt/tsconfig.bazel.json
@@ -7,46 +7,20 @@
     "strict": true,
     "skipLibCheck": true,
     "noEmit": true,
-    "types": [
-      "node",
-      "react",
-      "react-dom",
-      "@vitest/browser-playwright"
-    ],
-    "lib": [
-      "ES2022",
-      "DOM",
-      "DOM.Iterable"
-    ],
+    "types": ["node", "react", "react-dom"],
+    "lib": ["ES2022", "DOM", "DOM.Iterable"],
     "paths": {
-      "#packages/editor/*": [
-        "../*"
-      ],
-      "react": [
-        "./node_modules/@types/react"
-      ],
-      "react-dom": [
-        "./node_modules/@types/react-dom"
-      ],
-      "react/jsx-runtime": [
-        "./node_modules/@types/react/jsx-runtime"
-      ],
-      "@material-ui/*": [
-        "./node_modules/@material-ui/*"
-      ],
-      "react-intl": [
-        "./node_modules/react-intl"
-      ],
+      "#packages/editor/*": ["../*"],
+      "react": ["./node_modules/@types/react"],
+      "react-dom": ["./node_modules/@types/react-dom"],
+      "react/jsx-runtime": ["./node_modules/@types/react/jsx-runtime"],
+      "@material-ui/*": ["./node_modules/@material-ui/*"],
+      "react-intl": ["./node_modules/react-intl"],
       "@formatjs/icu-messageformat-parser": [
         "./node_modules/@formatjs/icu-messageformat-parser"
       ]
     }
   },
-  "include": [
-    "*.ts",
-    "*.tsx",
-    "../*.ts",
-    "../*.tsx"
-  ],
+  "include": ["*.ts", "*.tsx", "../*.ts", "../*.tsx"],
   "exclude": []
 }
